Add Container Insights/Azure Monitor backend and demo workloads
- InsightsMetricsRepository: read K8S container CPU/memory from Log
  Analytics Perf table, joined to KubePodInventory per deployment
- AzureMetricsRepository with service-principal auth
- Support fractional (sub-day) duration ranges in controllers/service
- Mock repo emits 60s-granularity points for ranges under 1 hour
- Add insights/azure profiles, .env.example, README setup docs
- Add azure-demo (nginx busy/idle workloads + traffic scripts)

+++ src/main/java/com/resourceautoscaler/controller/MetricsController.java
@@ -27,7 +27,7 @@ public class MetricsController {
     @GetMapping("/{resourceId}")
     public ResponseEntity<MetricsResponse> getResourceMetrics(
             @PathVariable String resourceId,
-            @RequestParam(defaultValue = "30") int days
+            @RequestParam(defaultValue = "30") double days
     ) {
         ResourceMetrics metrics = metricsService.collectMetrics(resourceId, days);
         PeakHoursConfig config = new PeakHoursConfig(
